requirements.txt
semantic-kernel>=1.0.0
openai>=1.0.0
ddgs
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
python-docx>=1.1.0
streamlit>=1.30.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
